refactor: lazy load components
