+++ src/components/mobile/mobile.vue
@@ -84,7 +84,6 @@
   import { Component } from 'vue-property-decorator';
   import Main from '../main.vue';
   import ScrollToTop from '../scroll-to-top.vue';
-  import Sort from '../sort.vue';
   import Empathize from '../predictive-layer.vue';
   import SearchBox from '../search-box.vue';
   import HasSearchedMixin from '../has-searched.mixin.vue';
@@ -109,7 +108,7 @@
       ScrollToTop,
       SearchBox,
       SingleColumnLayout,
-      Sort,
+      Sort: () => import('../search').then(m => m.Sort),
       RelatedTags: () => import('../search').then(m => m.RelatedTags),
       SelectedFilters: () => import('../search').then(m => m.SelectedFilters),
       Facets: () => import('../search').then(m => m.Facets)
